requirements.txt
fastapi
uvicorn
python-json-logger
structlog
pydantic_settings


structlog
requests
lxml
jinja2


websockets
git+https://github.com/dhan-oss/DhanHQ-py.git
python-multipart
pandas-stubs

apscheduler
websocket-client
